Subscribers cannot delete themselves (if they're currently logged in).
git-svn-id: svn+ssh://cvs1/svn/listen/trunk@992 4692729f-32a0-46b7-9832-1bf53c4881b5

+++ controller/src/main/java/com/interact/listen/gui/DeleteSubscriberServlet.java
@@ -58,6 +58,11 @@ public class DeleteSubscriberServlet extends HttpServlet
             throw new UnauthorizedServletException();
         }
 
+        if(subscriberToDelete.equals(currentSubscriber))
+        {
+            throw new UnauthorizedServletException();
+        }
+
         PersistenceService persistenceService = new PersistenceService(session, currentSubscriber, Channel.GUI);
         History.deleteAllBySubscriber(session, subscriberToDelete);
         persistenceService.delete(subscriberToDelete);
